Setup e2e tests

+++ apps/server/src/user/user.controller.ts
@@ -67,6 +67,7 @@ export class UserController {
   @HttpCode(HttpStatus.NO_CONTENT)
   @Delete(':id')
   remove(@Param('id') id: string): void {
+    if (!Number.isInteger(Number(id))) throw new HttpException('Invalid user id', HttpStatus.BAD_REQUEST);
     this.userService.delete({ id: Number(id) });
   }
 }
